Improved GraphQL server integration
